Resolved PR
